프로젝트 수정

+++ api/src/main/java/com/bithumbsystems/lrc/management/api/v1/lrcmanagment/foundation/model/response/MarketResponse.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class MarketResponse {
 
-    private Long MinimumQuantity;
-    private Long actualQuantity;
+    private Double MinimumQuantity;
+    private Double actualQuantity;
     private String symbol;
 }
