+++ api/src/main/java/com/bithumbsystems/lrc/management/api/v1/lrcmanagment/project/icoinfo/model/request/IcoInfoRequest.java
@@ -29,7 +29,7 @@ public class IcoInfoRequest {
         private String marketInfo;          //마켓 정보
 
         @Schema(description = "상장가(원)")
-        private Long price;                 //상장가(원)
+        private Double price;                 //상장가(원)
 
         @Schema(description = "상장일")
         private LocalDate icoDate;      //상장일
