+++ api/src/main/java/com/bithumbsystems/lrc/management/api/v1/lrcmanagment/project/marketingquantity/model/response/MarketingQuantityResponse.java
@@ -12,6 +12,6 @@ public class MarketingQuantityResponse {
     private String id;
     private String projectId;           //프로젝트 id
     private String symbol;              //심볼
-    private Long minimumQuantity;     //최소 지원 수량
-    private Long actualQuantity;      //실제 상장 지원 수량
+    private Double minimumQuantity;     //최소 지원 수량
+    private Double actualQuantity;      //실제 상장 지원 수량
 }
